Fix CUDA NaN asserts in bfloat16 generation by floating logits

+++ vvembed/modular/modeling_vibevoice.py
@@ -433,7 +433,7 @@ class VibeVoiceForConditionalGeneration(VibeVoicePreTrainedModel):
 
         hidden_states = outputs.last_hidden_state
         logits = self.lm_head(hidden_states)
-        # logits = logits.float()
+        logits = logits.float()
 
         loss = None
         if labels is not None:
